scripts/requirements.txt
autopep8==1.5.4
boto3==1.16.23
botocore==1.19.23
certifi==2020.11.8
chardet==3.0.4
docker==4.3.1
environs==9.2.0
idna==2.10
jmespath==0.10.0
marshmallow==3.9.1
pycodestyle==2.6.0
python-dateutil==2.8.1
python-dotenv==0.15.0
requests==2.25.0
s3transfer==0.3.3
six==1.15.0
toml==0.10.2
urllib3==1.26.5
websocket-client==0.57.0
psycopg2-binary